feat: Add Wanted page and update branding
Introduces a new "Wanted" page for the Sk8Hunt, accessible via a new route and navigation link. This change also includes an update to the application's branding and hero section on the donation page to reflect the "JayNdaboX.IcU" theme.

Additionally, a new QR code for Venmo donations has been added to constants. The Orbitron font has been integrated for heading elements, and the `gps_enabled` property has been added to the user type.